feat/study_domain (#4)
* feat: resolve merge conflicts

* feat: product repo and service added

* feat: preparation repo and service added

* feat: security init

* feat: cleanup ResponseEntity

* feat: UserInfo init

* feat: UserInfo init + made code more conform to DDD

* feat: security added to UserInfo and Product

* feat: Consumption logic added

* feat: init integration tests

* feat: integration tests added

* feat: optimization of the domain definition

* feat: addition of Study + exceptions cleanup

* feat: filter added

---------

+++ src/main/kotlin/be/ehb/gdt/nutrisearch/domain/product/valueobjects/ServingSize.kt
@@ -1,8 +1,8 @@
 package be.ehb.gdt.nutrisearch.domain.product.valueobjects
 
-class ServingSize(val grams: Double, val name: String) {
+class ServingSize(val grams: Int, val name: String) {
 
-    constructor() : this(1.0, "Gram")
+    constructor() : this(1, "Gram")
 
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
